+++ packages/@magic-ext/cosmos/src/type.ts
@@ -5,6 +5,7 @@ export interface CosmosConfig {
 
 export enum CosmosPayloadMethod {
   Sign = 'cos_sign',
+  SignTypedData = 'cos_signTypedData',
   SignAndBroadcast = 'cos_signAndBroadcast',
   SendTokens = 'cos_sendTokens',
   ChangeAddress = 'cos_changeAddress',